+++ R/tabulate.R
@@ -103,7 +103,7 @@ ypr_detabulate_parameters <- function(x) {
 #' Table Fish
 #'
 #' Produces a data frame of the number of fish in the
-#' 'Surviving', 'Spawning', 'Caught', 'Harvested', 'Released' and 'HandlingMortality' categories by
+#' 'Surviving', 'Spawners', 'Caught', 'Harvested', 'Released' and 'HandlingMortality' categories by
 #' 'Length', 'Age' or 'Weight' class.
 #'
 #' @inheritParams params
@@ -124,13 +124,13 @@ ypr_tabulate_fish <- function(population, x = "Age", binwidth = 1L) {
   table <- as.data.frame(table)
 
   table$Surviving <- table$FishedSurvivorship * sr(table, population)$R0F
-  table$Spawning <- table$Surviving * table$Spawning
+  table$Spawners <- table$Surviving * table$Spawning
   table$Caught <- table$Surviving *  table$Vulnerability * population$pi
   table$Harvested <- table$Caught * table$Retention
   table$Released <- table$Caught * (1 - table$Retention)
   table$HandlingMortality <- table$Released * population$Hm
 
-  table <- table[c(x, "Surviving", "Spawning", "Caught", "Harvested",
+  table <- table[c(x, "Surviving", "Spawners", "Caught", "Harvested",
     "Released", "HandlingMortality")]
 
   breaks <- seq(0, max(table[[1]] + binwidth), by = binwidth)