+++ tests/testthat/test-tabulate.R
@@ -93,13 +93,13 @@ test_that("ypr_tabulate_yield", {
   expect_identical(sr$Type, c("unfished", "actual", "optimal"))
 
   fish <- ypr_tabulate_fish(ypr_population())
-  expect_identical(colnames(fish), c("Age", "Surviving", "Spawning", "Caught",
+  expect_identical(colnames(fish), c("Age", "Surviving", "Spawners", "Caught",
     "Harvested", "Released", "HandlingMortality"))
   expect_identical(fish[[1]], as.double(1:20))
   expect_equal(fish$Surviving[1:2], c(0.134, 0.110), tolerance = 0.001)
 
   fish <- ypr_tabulate_fish(ypr_population(), x = "Length")
-  expect_identical(colnames(fish), c("Length", "Surviving", "Spawning", "Caught",
+  expect_identical(colnames(fish), c("Length", "Surviving", "Spawners", "Caught",
     "Harvested", "Released", "HandlingMortality"))
   expect_identical(fish$Length[1:2], c(14, 26))